running-coach-mvp/requirements.txt
fastapi
uvicorn
sqlalchemy
stravalib
numpy
scipy
pandas
filterpy
transformers
torch
pytest
click
python-dotenv
httpx
